Full Stack Open Part 2e: Exercise 2.19: countries - Add show button to SearchListings component. Create and implement handleShowCountry click handler. Set searchString to country.name.common on click. Create Button component. Refactor CountryInfoPanel and sub components into CountryInfoPanel.jsx. Rename SearchResults to CountrySearch module component feature
