+++ exercises/parts-2-a--2-e/countries/src/components/Button.jsx
@@ -1,0 +1,9 @@
+const Button = ({label, type, onClickHandler}) => {
+    return (
+        <>
+            <button onClick={onClickHandler} type={type}>{label}</button>
+        </>
+    )
+}
+
+export default Button
